+++ Realm/RLMRealm.h
@@ -18,7 +18,7 @@
 
 #import <Foundation/Foundation.h>
 #import "RLMConstants.h"
-#import "RLMHandover.h"
+#import "RLMThreadSafeReference.h"
 
 @class RLMRealmConfiguration, RLMObject, RLMSchema, RLMMigration, RLMNotificationToken;
 
@@ -339,6 +339,8 @@ typedef void (^RLMNotificationBlock)(RLMNotification notification, RLMRealm *rea
 
 #pragma mark - Accessing Objects
 
+// TODO: Document
+- (nullable id)resolveThreadSafeReference:(RLMThreadSafeReference *)reference;
 
 #pragma mark - Adding and Removing Objects from a Realm
 
@@ -495,22 +497,6 @@ __deprecated_msg("Use `performMigrationForConfiguration:error:`") NS_REFINED_FOR
  @see                 RLMMigration
  */
 + (BOOL)performMigrationForConfiguration:(RLMRealmConfiguration *)configuration error:(NSError **)error;
-
-/**
- Creates a `RLMThreadHandover` from the passed in objects that can be safely imported on a separate thread.
-
- Used to pass objects between threads in a thread-safe manner.
-
- @warning Note that the returned `ThreadHandover` must be imported at most once, and that the current version of the
- Realm will remain pinned until this package is imported or deallocated.
-
- @param objects The objects to pass between threads.
- @return A package that can be imported on a separate thread.
-
- @see RLMThreadHandover
- */
-- (RLMThreadHandover *)exportThreadHandoverWithObjects:(NSArray<id<RLMThreadConfined>> *)objects
-NS_SWIFT_NAME(exportThreadHandover(containing:));
 
 @end