+++ Realm/RLMRealm.mm
@@ -32,8 +32,8 @@
 #import "RLMQueryUtil.hpp"
 #import "RLMRealmUtil.hpp"
 #import "RLMSchema_Private.hpp"
+#import "RLMThreadSafeReference_Private.hpp"
 #import "RLMUpdateChecker.hpp"
-#import "RLMHandover_Private.hpp"
 #import "RLMUtil.hpp"
 
 #include "impl/realm_coordinator.hpp"
@@ -471,6 +471,11 @@ REALM_NOINLINE void RLMRealmTranslateException(NSError **error) {
     }
 }
 
+- (nullable id)resolveThreadSafeReference:(RLMThreadSafeReference *)reference {
+    id<RLMThreadConfined> threadConfined = [reference resolveReferenceInRealm:self];
+    return threadConfined.invalidated ? nil : threadConfined;
+}
+
 /**
  Replaces all string columns in this Realm with a string enumeration column and compacts the
  database file.
@@ -642,10 +647,6 @@ REALM_NOINLINE void RLMRealmTranslateException(NSError **error) {
         *error = localError; // Must set outside pool otherwise will free anyway
     }
     return success;
-}
-
-- (RLMThreadHandover *)exportThreadHandoverWithObjects:(NSArray<id<RLMThreadConfined>> *)objects {
-    return [[RLMThreadHandover alloc] initWithRealm:self objects:objects];
 }
 
 - (RLMObject *)createObject:(NSString *)className withValue:(id)value {