+++ RealmSwift/List.swift
@@ -51,8 +51,7 @@ Lists can be filtered and sorted with the same predicates as `Results<T>`.
 
 When added as a property on `Object` models, the property must be declared as `let` and cannot be `dynamic`.
 */
-public final class List<T: Object>: ListBase, ThreadConfined {
-    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
+public final class List<T: Object>: ListBase {
 
     /// Element type contained in this collection.
     public typealias Element = T
@@ -69,10 +68,6 @@ public final class List<T: Object>: ListBase, ThreadConfined {
     public var isInvalidated: Bool { return _rlmArray.isInvalidated }
 
     // MARK: Initializers
-
-    fileprivate override init!(array: RLMArray<RLMObject>!) {
-        super.init(array: array)
-    }
 
     /// Creates a `List` that holds objects of type `T`.
     public override init() {
@@ -480,7 +475,7 @@ public final class List<T: Object>: ListBase, ThreadConfined {
     }
 }
 
-extension List: RealmCollection, RangeReplaceableCollection {
+extension List : RealmCollection, RangeReplaceableCollection {
     // MARK: Sequence Support
 
     /// Returns a `RLMIterator` that yields successive elements in the `List`.
@@ -579,20 +574,6 @@ extension List {
     public func average<U: AddableType>(_ property: String) -> U? { fatalError() }
 }
 
-extension List: _ThreadConfined {
-    var bridgedData: RLMThreadConfined {
-        return _rlmArray
-    }
-
-    var bridgedMetadata: Any? {
-        return nil
-    }
-
-    static func bridge(data: RLMThreadConfined, metadata: Any?) -> List {
-        return List(array: data as! RLMArray)
-    }
-}
-
 #else
 
 /// :nodoc:
@@ -627,8 +608,7 @@ public class ListBase: RLMListBase {
 
  Properties of `List` type defined on `Object` subclasses must be declared as `let` and cannot be `dynamic`.
 */
-public final class List<T: Object>: ListBase, ThreadConfined {
-    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
+public final class List<T: Object>: ListBase {
 
     /// The type of the elements contained within the collection.
     public typealias Element = T
@@ -644,10 +624,6 @@ public final class List<T: Object>: ListBase, ThreadConfined {
     public var invalidated: Bool { return _rlmArray.invalidated }
 
     // MARK: Initializers
-
-    private override init!(array: RLMArray) {
-        super.init(array: array)
-    }
 
     /// Creates a `List` that holds Realm model objects of type `T`.
     public override init() {
@@ -1090,20 +1066,6 @@ extension List: RealmCollectionType, RangeReplaceableCollectionType {
         return _rlmArray.addNotificationBlock { _, change, error in
             block(RealmCollectionChange.fromObjc(anyCollection, change: change, error: error))
         }
-    }
-}
-
-extension List: _ThreadConfined {
-    var bridgedData: RLMThreadConfined {
-        return _rlmArray
-    }
-
-    var bridgedMetadata: Any? {
-        return nil
-    }
-
-    static func bridge(data: RLMThreadConfined, metadata: Any?) -> List {
-        return List(array: data as! RLMArray)
     }
 }