+++ RealmSwift/Results.swift
@@ -72,8 +72,7 @@ up-to-date done on a background thread whenever possible.
 
 Results cannot be created directly.
 */
-public final class Results<T: Object>: NSObject, NSFastEnumeration, ThreadConfined {
-    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
+public final class Results<T: Object>: NSObject, NSFastEnumeration {
 
     internal let rlmResults: RLMResults<RLMObject>
 
@@ -383,22 +382,6 @@ public final class Results<T: Object>: NSObject, NSFastEnumeration, ThreadConfin
     }
 }
 
-// MARK: ThreadConfined
-
-extension Results: _ThreadConfined {
-    var bridgedData: RLMThreadConfined {
-        return rlmResults
-    }
-
-    var bridgedMetadata: Any? {
-        return nil
-    }
-
-    static func bridge(data: RLMThreadConfined, metadata: Any?) -> Results {
-        return Results(data as! RLMResults)
-    }
-}
-
 extension Results: RealmCollection {
     // MARK: Sequence Support
 
@@ -559,8 +542,7 @@ public class ResultsBase: NSObject, NSFastEnumeration {
 
  `Results` cannot be directly instantiated.
 */
-public final class Results<T: Object>: ResultsBase, ThreadConfined {
-    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
+public final class Results<T: Object>: ResultsBase {
 
     /// The type of the objects contained in the collection.
     public typealias Element = T
@@ -847,7 +829,7 @@ public final class Results<T: Object>: ResultsBase, ThreadConfined {
     }
 }
 
-    extension Results: RealmCollectionType {
+extension Results: RealmCollectionType {
     // MARK: Sequence Support
 
     /// Returns an `RLMGenerator` that yields successive elements in the results.
@@ -873,22 +855,6 @@ public final class Results<T: Object>: ResultsBase, ThreadConfined {
         return rlmResults.addNotificationBlock { _, change, error in
             block(RealmCollectionChange.fromObjc(anyCollection, change: change, error: error))
         }
-    }
-}
-
-// MARK: ThreadConfined
-
-extension Results: _ThreadConfined {
-    var bridgedData: RLMThreadConfined {
-        return rlmResults
-    }
-
-    var bridgedMetadata: Any? {
-        return nil
-    }
-
-    static func bridge(data: RLMThreadConfined, metadata: Any?) -> Results {
-        return Results(data as! RLMResults)
     }
 }