+++ RealmSwift/Tests/TestCase.swift
@@ -106,7 +106,7 @@ class TestCase: XCTestCase {
         RLMRealm.resetRealmState()
     }
 
-    func dispatchAsyncAndWait(block: @escaping () -> ()) {
+    func dispatchSyncNewThread(block: @escaping () -> Void) {
         queue.async {
             autoreleasepool {
                 block()
@@ -185,8 +185,6 @@ class TestCase: XCTestCase {
     var exceptionThrown = false
     var testDir: String! = nil
 
-    let queue = dispatch_queue_create("background", nil)
-
     func realmWithTestPath(configuration: Realm.Configuration = Realm.Configuration()) -> Realm {
         var configuration = configuration
         configuration.fileURL = testRealmURL()
@@ -255,8 +253,9 @@ class TestCase: XCTestCase {
     func resetRealmState() {
         RLMRealm.resetRealmState()
     }
-    
-    func dispatchAsyncAndWait(block: dispatch_block_t) {
+
+    func dispatchSyncNewThread(block: dispatch_block_t) {
+        let queue = dispatch_queue_create("background", nil)
         dispatch_async(queue) {
             autoreleasepool {
                 block()