+++ CHANGELOG.md
@@ -12,13 +12,13 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### Enhancements
 
-* Introduce API for safely handing objects between threads.
 * Improve error message when using NSNumber incorrectly in Swift models.
 * Fix compilation with Swift 2.3 using Xcode 8 beta 2.
 * Further reduce the download size of the prebuilt static libraries.
 * Improve sort performance, especially on non-nullable columns.
 * Allow partial initialization of object by `initWithValue:`, deferring
   required property checks until object is added to Realm.
+* Introduce API for safely handing objects between threads.
 
 ### Bugfixes